eval supports multiple evaluation

+++ tdmpc2/train.py
@@ -62,7 +62,7 @@ def train(cfg: dict):
     set_seed(cfg.seed)
     print(colored("Work dir:", "yellow", attrs=["bold"]), cfg.work_dir)
 
-    if cfg.overwrite:
+    if cfg.overwrite and os.path.exists(cfg.work_dir):
         shutil.rmtree(cfg.work_dir)
 
     trainer_cls = OfflineTrainer if cfg.multitask else OnlineTrainer
